4776 - Clean up user store hooks (auth related)
