+++ src/client/pages/Country/SectionWrapper/hooks/useReviewStatusListener.tsx
@@ -6,7 +6,7 @@ import { Sockets } from 'meta/socket'
 import { useAppDispatch } from 'client/store/hooks'
 import { useOriginalDataPoint } from 'client/store/ui/originalDataPoint'
 import { ReviewActions } from 'client/store/ui/review'
-import { useCanEditCycleData } from 'client/store/user'
+import { useCanEditCycleData } from 'client/store/user/hooks/auth'
 import { useUser } from 'client/store/user/hooks/user'
 import { useSectionRouteParams } from 'client/hooks/useRouteParams'
 import { SocketClient } from 'client/service/socket'
