+++ src/client/pages/Country/hooks/useReviewSummaryListener.tsx
@@ -4,7 +4,7 @@ import { Sockets } from 'meta/socket'
 
 import { useAppDispatch } from 'client/store/hooks'
 import { ReviewActions } from 'client/store/ui/review'
-import { useCanEditCycleData } from 'client/store/user'
+import { useCanEditCycleData } from 'client/store/user/hooks/auth'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
 import { SocketClient } from 'client/service/socket'
