+++ src/client/pages/CountryHome/Collaborators/UserList/UserCard/Actions/useActions.ts
@@ -5,8 +5,7 @@ import { Authorizer, UserInvitations, Users } from 'meta/user'
 import { CountryUserSummaries } from 'meta/user/countryUserSummaries'
 
 import { useCycle } from 'client/store/meta/hooks/cycles'
-import { useCanSeeUserActivities } from 'client/store/user'
-import { useCanEditUserActivities } from 'client/store/user/hooks'
+import { useCanEditUserActivities, useCanSeeUserActivities } from 'client/store/user/hooks/auth'
 import { useUser } from 'client/store/user/hooks/user'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
