+++ src/client/pages/OriginalDataPoint/OriginalDataPoint.tsx
@@ -7,7 +7,7 @@ import { useNavigate } from 'react-router-dom'
 import { TableNames } from 'meta/assessment/table'
 
 import { useOriginalDataPoint } from 'client/store/ui/originalDataPoint'
-import { useIsEditTableDataEnabled } from 'client/store/user'
+import { useIsEditTableDataEnabled } from 'client/store/user/hooks/auth'
 import { useCountryIso } from 'client/hooks'
 
 import ButtonBar from './components/ButtonBar'
