+++ src/client/pages/OriginalDataPoint/components/YearSelection/YearSelection.tsx
@@ -6,7 +6,7 @@ import classNames from 'classnames'
 import { ODPs } from 'meta/assessment/odps'
 
 import { useODPYears, useOriginalDataPoint } from 'client/store/ui/originalDataPoint'
-import { useIsEditTableDataEnabled } from 'client/store/user'
+import { useIsEditTableDataEnabled } from 'client/store/user/hooks/auth'
 import { useOriginalDataPointRouteParams } from 'client/hooks/useRouteParams'
 
 import { useOnChange } from './hooks/useOnChange'
