+++ src/client/pages/OriginalDataPoint/hooks/useShowReviewIndicator.ts
@@ -1,6 +1,6 @@
 import { SectionName, SectionNames } from 'meta/assessment/section'
 
-import { useCanViewReview } from 'client/store/user/hooks'
+import { useCanViewReview } from 'client/store/user/hooks/auth'
 import { useIsEditODPEnabled } from 'client/pages/OriginalDataPoint/hooks/useIsEditODPEnabled'
 
 export const useShowReviewIndicator = (sectionName: SectionName = SectionNames.extentOfForest) => {
