+++ src/client/pages/Section/Contacts/ContactRow/ContactRow.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { Contact } from 'meta/cycleData'
 
-import { useIsEditTableDataEnabled } from 'client/store/user'
+import { useIsEditTableDataEnabled } from 'client/store/user/hooks/auth'
 import { DataCell, DataRow } from 'client/components/DataGrid'
 import CellNodeExt from 'client/components/TableNodeExt/CellNodeExt'
 import { Columns, Fields } from 'client/pages/Section/Contacts/types'
