+++ src/client/pages/Section/Contacts/hooks/useContactsData.ts
@@ -7,7 +7,7 @@ import { Contact, ContactField, Contacts } from 'meta/cycleData'
 import { RoleName } from 'meta/user'
 
 import { useContacts } from 'client/store/data'
-import { useIsEditTableDataEnabled } from 'client/store/user'
+import { useIsEditTableDataEnabled } from 'client/store/user/hooks/auth'
 import { useIsPrintRoute } from 'client/hooks/useIsRoute'
 import { useSectionContext } from 'client/pages/Section/context'
