+++ src/client/pages/Section/DataTable/DataTable.tsx
@@ -7,7 +7,7 @@ import { Table as TableType } from 'meta/assessment/table'
 import { RecordAssessmentDatas } from 'meta/data'
 
 import { useCycle } from 'client/store/meta/hooks/cycles'
-import { useIsEditTableDataEnabled } from 'client/store/user'
+import { useIsEditTableDataEnabled } from 'client/store/user/hooks/auth'
 import { useIsPrintRoute } from 'client/hooks/useIsRoute'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
