+++ src/client/pages/Section/DataTable/Table/RowData/hooks/useRowActions.tsx
@@ -8,7 +8,7 @@ import { Table } from 'meta/assessment/table'
 import { Topics } from 'meta/messageCenter'
 
 import { useCycle } from 'client/store/meta/hooks/cycles'
-import { useCanViewReview } from 'client/store/user/hooks'
+import { useCanViewReview } from 'client/store/user/hooks/auth'
 import { DataRowAction, DataRowActionType } from 'client/components/DataGrid'
 
 type Props = {
