+++ src/client/pages/Section/DataTable/Table/Table.tsx
@@ -8,8 +8,7 @@ import { Table as TableType } from 'meta/assessment/table'
 import { RecordAssessmentData } from 'meta/data'
 
 import { useIsDataLocked } from 'client/store/ui/dataLock'
-import { useCanEdit } from 'client/store/user'
-import { useCanViewReview } from 'client/store/user/hooks'
+import { useCanEdit, useCanViewReview } from 'client/store/user/hooks/auth'
 import { useIsPrintRoute } from 'client/hooks/useIsRoute'
 import { ButtonGridExport, DataGrid } from 'client/components/DataGrid'
 import ButtonCopyValues from 'client/pages/Section/DataTable/Table/ButtonCopyValues'
