+++ src/client/pages/Section/Descriptions/CommentableDescription/Title/hooks/useDescriptionActions.ts
@@ -4,7 +4,7 @@ import { CountryIso } from 'meta/area'
 import { CommentableDescriptionName } from 'meta/assessment/descriptionValue'
 import { Topics } from 'meta/messageCenter'
 
-import { useCanEditDescription } from 'client/store/user/hooks'
+import { useCanEditDescription } from 'client/store/user/hooks/auth'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
 import { DataRowAction, DataRowActionType } from 'client/components/DataGrid'
 import { useSectionContext } from 'client/pages/Section/context'
