+++ src/client/pages/Section/Descriptions/NationalDataDescriptions/DataSources/hooks/useDataSourcesData.ts
@@ -5,7 +5,7 @@ import { SectionName } from 'meta/assessment/section'
 import { UUIDs } from 'meta/uuid'
 
 import { useCommentableDescriptionValue } from 'client/store/data'
-import { useIsDescriptionEditable } from 'client/store/user/hooks'
+import { useIsDescriptionEditable } from 'client/store/user/hooks/auth'
 
 type Props = {
   sectionName: SectionName
