+++ src/client/pages/Section/Section.tsx
@@ -11,7 +11,7 @@ import { useIsSectionDataEmpty } from 'client/store/data'
 import { useCycle } from 'client/store/meta/hooks/cycles'
 import { useSection } from 'client/store/meta/hooks/sections'
 import { useTableSections } from 'client/store/meta/hooks/tableSections'
-import { useIsEditTableDataEnabled } from 'client/store/user/hooks'
+import { useIsEditTableDataEnabled } from 'client/store/user/hooks/auth'
 import { useIsPrintRoute } from 'client/hooks/useIsRoute'
 import { useSectionRouteParams } from 'client/hooks/useRouteParams'
 import { SectionContext, SectionContextValue } from 'client/pages/Section/context'
