+++ src/client/pages/Section/SectionHeader/ForestCharacteristics/ForestCharacteristics.tsx
@@ -7,7 +7,7 @@ import { AreaActions } from 'client/store/area/actions'
 import { useCountry, useIsUpdatingCountry } from 'client/store/area/hooks/country'
 import { useHasOriginalDataPointData } from 'client/store/data'
 import { useAppDispatch } from 'client/store/hooks'
-import { useIsEditTableDataEnabled } from 'client/store/user'
+import { useIsEditTableDataEnabled } from 'client/store/user/hooks/auth'
 import { useUser } from 'client/store/user/hooks/user'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
 import Button, { ButtonSize } from 'client/components/Buttons/Button'
